requirements.txt
crewai[google-genai,groq]>=0.83.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
litellm>=1.59.0
crewai-tools>=0.12.0